requirements.txt
alt_profanity_check==1.0.1
Flask==2.2.3
Flask_Cors==3.0.10
pydub
tqdm==4.65.0
wenetruntime